src/requirements.txt
django==2.1
django-environ==0.4.5
